Allow hooks to suppress messages
By returning a canned error that's exported in the logrus package, hooks
can choose to suppress messages entirely.

This creates two considerations:

- If your hooks are publishing to some upstream service, the order in
  which hooks are added matters. An upstream hook can squelch a downstream
  hook.

- Looking at this from the opposite perspective, hooks upstream of the
  one that choose not to suppress will be fully fired before a
  downstream hook which chooses to squelch the message, suppressing the
  overall output as well as further downstream hooks.

In general, this means that the best practice here is that if you have a
hook that may suppress a log message, you should add it as early as
possible, ahead of the hooks that don't care about suppressing messages.